// components/admin/adminUsers/adminUsersEditor/adminUsersEditorStyle.scss.d.ts
export type Styles = {
  'categoryContentContainer': string;
  'categorySeparator': string;
  'editorIsAllHeadsListContainer': string;
  'errorText': string;
  'inputText': string;
  'popupBodySize': string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;